requirements.txt
Django==5.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Pillow==10.2.0
python-dotenv==1.0.1
drf-yasg==1.21.7
setuptools>=68.0.0 